+++ crates/voicevox_core_macros/src/mora_mappings.rs
@@ -52,8 +52,8 @@ pub(crate) fn derive_mora_mappings(input: &DeriveInput) -> syn::Result<proc_macr
         let values = variants.iter().map(|(_, _, consonant, vowel)| {
             quote! {
                 (
-                    crate::engine::acoustic_feature_extractor::optional_consonant!(#consonant),
-                    crate::engine::acoustic_feature_extractor::mora_tail!(#vowel)
+                    crate::engine::acoustic_feature_extractor::convert::optional_consonant!(#consonant),
+                    crate::engine::acoustic_feature_extractor::convert::mora_tail!(#vowel)
                 )
             }
         });
